+++ src/main/java/com/farmdora/farmdora/opinion/dto/ProcessType.java
@@ -1,4 +1,4 @@
-package com.farmdora.farmdora.question.dto;
+package com.farmdora.farmdora.opinion.dto;
 
 public enum ProcessType {
     WAIT, DONE
