+++ src/main/java/com/farmdora/farmdora/opinion/service/OpinionService.java
@@ -1,9 +1,9 @@
-package com.farmdora.farmdora.question.service;
+package com.farmdora.farmdora.opinion.service;
 
 import com.farmdora.farmdora.common.response.PageResponseDto;
-import com.farmdora.farmdora.question.dto.QuestionResponseDto;
-import com.farmdora.farmdora.question.dto.QuestionSearchRequestDto;
-import com.farmdora.farmdora.question.repository.QuestionRepository;
+import com.farmdora.farmdora.opinion.dto.QuestionResponseDto;
+import com.farmdora.farmdora.opinion.dto.QuestionSearchRequestDto;
+import com.farmdora.farmdora.opinion.repository.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class QuestionService {
+public class OpinionService {
     private final QuestionRepository questionRepository;
 
     @Transactional(readOnly = true)
