+++ src/test/java/com/farmdora/farmdora/opinion/controller/OpinionControllerTest.java
@@ -1,4 +1,4 @@
-package com.farmdora.farmdora.question.controller;
+package com.farmdora.farmdora.opinion.controller;
 
 import static com.farmdora.farmdora.common.response.SuccessMessage.SEARCH_QUESTION_SUCCESS;
 import static org.hamcrest.core.IsEqual.equalTo;
@@ -11,8 +11,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.farmdora.farmdora.ControllerTest;
 import com.farmdora.farmdora.common.response.PageResponseDto;
-import com.farmdora.farmdora.question.dto.QuestionResponseDto;
-import com.farmdora.farmdora.question.dto.QuestionSearchRequestDto;
+import com.farmdora.farmdora.opinion.dto.QuestionResponseDto;
+import com.farmdora.farmdora.opinion.dto.QuestionSearchRequestDto;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -22,7 +22,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
-class QuestionControllerTest extends ControllerTest {
+class OpinionControllerTest extends ControllerTest {
 
     @Test
     @DisplayName("문의 목록 검색 API 테스트")
@@ -49,7 +49,7 @@ class QuestionControllerTest extends ControllerTest {
         Pageable pageable = PageRequest.of(0, 10);
         Page<QuestionResponseDto> questionPage = new PageImpl<>(questions, pageable, 2);
         PageResponseDto<QuestionResponseDto> result = new PageResponseDto<>(questions, questionPage);
-        when(questionService.searchQuestions(anyInt(), any(QuestionSearchRequestDto.class), any(Pageable.class))).thenReturn(result);
+        when(opinionService.searchQuestions(anyInt(), any(QuestionSearchRequestDto.class), any(Pageable.class))).thenReturn(result);
 
         // when
         // then
